--- packages/common/en.ts
@@ -1,10 +1,0 @@
-export const ERROR_MESSAGES = {
-  questionController: {
-    createQuestion: {
-      invalidQueue: "Posted to an invalid queue",
-      notInQueue: "Can't post question to course you're not in!",
-      closedQueue: "You can't post a question to a closed queue",
-      multipleQuestions: "You can't create more than one question at a time.",
-    },
-  },
-};
